Add files via upload
Updated iteration of Bookish Flexdash with the option for genre tabs. To utilise these, you will need to change the names of the tabs and the calls for R to subset the data by genre. The key for this will generally to be to watch for (and then change!) calls for (Genre == "Genre_Name_Here"). 

It's very important that you're consistently recording these terms the same way in your reading log.  R is only as clever as what you tell it to do and it won't recognise misspellings or when you fail to capitalise something ("romance" vs "Romance"). If you're careful in your data management, though, this shouldn't cause you many issues. 

Example 1:
History {data-navmenu="Genres"}

For this tab, you would want to change the tab title "History" to something else that would be more suitable for your own reading habits. 

Example 2:
bigrams <- subset(TBR, (Genre == "History") & (Start_Finish == "1"))

The same action will need to be taken here to change (Genre == "History") to whatever you would like the tab to reflect

Example 3:
filter((!is.na(Month)) & (Start_Finish == "1") & (Genre == "History")) %>%

Again, you just need to swap out the "History" term for whatever will be more relevant for you